requirements.txt
arabert==1.0.1
fasttext==0.9.2
nltk==3.8.1
numpy==1.26.0
python_Levenshtein==0.22.0
scikit_learn==1.3.0
scipy==1.11.3
setuptools==67.8.0
transformers==4.33.2
